refactor(project-restructure): :art: Sync utils import
middleware
