+++ usagelogger/middleware/flask.py
@@ -9,7 +9,7 @@ from werkzeug.wrappers import Request
 from werkzeug.wsgi import ClosingIterator
 
 from usagelogger import HttpLogger, HttpMessage, HttpRequestImpl, HttpResponseImpl
-from usagelogger.multipart_utils import decode_multipart
+from usagelogger.utils.multipart_utils import decode_multipart
 
 
 class HttpLoggerForFlask:
